# docs/openapi.yaml
openapi: 3.1.0
info:
  title: AI Sales Platform API
  version: 2.0.0
  description: Authenticated, tenant-isolated CRM, reporting, settings, and AI workflows.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: /api
security:
  - bearerAuth: []
...
paths:
...
  /auth/register:
    post:
      operationId: registerAccount
      tags: [Authentication]
      security: []
      summary: Create an account and send a verification link
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterRequest" }
      responses:
        "201": { description: Account created; verification required }
        "429": { $ref: "#/components/responses/RateLimited" }
  /auth/verification/request:
    post:
      operationId: requestEmailVerification
      tags: [Authentication]
      security: []
      summary: Request a replacement verification link
      requestBody: { $ref: "#/components/requestBodies/EmailRequest" }
      responses:
        "202": { description: Generic accepted response prevents account enumeration }
        "429": { $ref: "#/components/responses/RateLimited" }
  /auth/verify-email:
    post:
      operationId: verifyEmail
      tags: [Authentication]
      security: []
      summary: Verify email and return one-time recovery codes
      requestBody: { $ref: "#/components/requestBodies/TokenRequest" }
      responses:
        "200": { description: Email verified and session established }
        "400": { $ref: "#/components/responses/BadRequest" }
  /auth/login:
    post:
      operationId: login
      tags: [Authentication]
      security: []
      summary: Sign in with a verified email and password
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200": { description: Access token returned and refresh cookie set }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /auth/refresh:
    post:
      operationId: refreshSession
      tags: [Authentication]
      security: [{ refreshCookie: [] }]
      summary: Atomically rotate a refresh session
      responses:
        "200": { description: New access token and refresh cookie returned }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/logout:
    post:
      operationId: logout
      tags: [Authentication]
      security: [{ refreshCookie: [] }]
      summary: Revoke the current refresh session
      responses:
        "204": { description: Session revoked or already absent }
        "429": { $ref: "#/components/responses/RateLimited" }
  /auth/password-reset/request:
    post:
      operationId: requestPasswordReset
      tags: [Authentication]
      security: []
      summary: Request a one-time password reset link
      requestBody: { $ref: "#/components/requestBodies/EmailRequest" }
      responses:
        "202": { description: Generic accepted response prevents account enumeration }
        "429": { $ref: "#/components/responses/RateLimited" }
  /auth/password-reset/confirm:
    post:
      operationId: confirmPasswordReset
      tags: [Authentication]
      security: []
      summary: Change password and revoke all sessions
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ResetPasswordRequest" }
      responses:
        "204": { description: Password changed and sessions revoked }
        "400": { $ref: "#/components/responses/BadRequest" }
  /auth/recover:
    post:
      operationId: recoverAccount
      tags: [Authentication]
      security: []
      summary: Recover an account with a one-time backup code
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RecoveryRequest" }
      responses:
        "204": { description: "Password changed, code consumed, and sessions revoked" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/recovery-codes:
    post:
      operationId: replaceRecoveryCodes
      tags: [Authentication]
      summary: Replace recovery codes after password confirmation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [password]
              properties: { password: { type: string, format: password, maxLength: 128 } }
      responses:
        "200": { description: New one-time recovery codes returned }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [Authentication]
      summary: Return the current user and settings
      responses:
        "200": { description: Current account }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    refreshCookie: { type: apiKey, in: cookie, name: refresh_token }
  schemas:
    Password: { type: string, format: password, minLength: 12, maxLength: 128 }
    RegisterRequest:
      type: object
      additionalProperties: false
      required: [name, email, password]
      properties:
        name: { type: string, minLength: 2, maxLength: 100 }
        email: { type: string, format: email, maxLength: 254 }
        password: { $ref: "#/components/schemas/Password" }
    LoginRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email: { type: string, format: email, maxLength: 254 }
        password: { type: string, format: password, maxLength: 128 }
    TokenRequest:
      type: object
      additionalProperties: false
      required: [token]
      properties: { token: { type: string, minLength: 32, maxLength: 256 } }
    ResetPasswordRequest:
      type: object
      additionalProperties: false
      required: [token, password]
      properties:
        token: { type: string, minLength: 32, maxLength: 256 }
        password: { $ref: "#/components/schemas/Password" }
    RecoveryRequest:
      type: object
      additionalProperties: false
      required: [email, recoveryCode, password]
      properties:
        email: { type: string, format: email, maxLength: 254 }
        recoveryCode: { type: string, minLength: 20, maxLength: 32 }
        password: { $ref: "#/components/schemas/Password" }
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
  requestBodies:
    EmailRequest:
      required: true
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            required: [email]
            properties: { email: { type: string, format: email, maxLength: 254 } }
    TokenRequest:
      required: true
      content:
        application/json:
          schema: { $ref: "#/components/schemas/TokenRequest" }
  responses:
    BadRequest: { description: Validation or token failure, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
    Unauthorized: { description: Authentication failed, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
    Forbidden: { description: Operation is not permitted, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
...
    RateLimited: { description: Request rate limit exceeded, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
...